+++ Web Quiz Engine with Java/task/src/engine/model/QuizOutputModel.java
@@ -2,5 +2,5 @@ package engine.model;
 
 import java.util.List;
 
-public record QuizOutputModel(Integer id, String title, String text, List<String> options) {
+public record QuizOutputModel(Long id, String title, String text, List<String> options) {
 }
